requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
pypdf==3.15.1
# Optional alternatives for PDF processing:
# pdfminer.six==20221105
# textract==1.6.5
